requirements.txt
jsonschema==4.21.1
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-mock==3.12.0
python-dotenv==1.0.0